Prompt 01
Results of the first prompt initialized `centipede_roguelike.py` with the game
code. In addition, Amazon Q also generated `requirements.txt` and `README.md`
files. The game runs without issues, but it's not to say the game itself didn't
have issues. For example, the graphics were sub-par, even for a retro game, and
the fire rate was much too high. Overall, however, it was an impressive start
out the gate.